Altered Piano insert function
To compensate for the lack of a menu driven system for inserts in the
higher level UI I have added extra checks in the insert functions for
all of the “piano metadata” tables to make sure that duplicatie entries
are not created but also allows for new entries to populate those
tables.

These functions are called on insert with every piano.